+++ experimental/packages/otlp-exporter-base/test/browser/xhr-transport.test.ts
@@ -26,11 +26,11 @@ import { ensureHeadersContain } from '../testHelper';
 
 const testTransportParameters = {
   url: 'http://example.test',
-  headers: {
+  headers: () => ({
     foo: 'foo-value',
     bar: 'bar-value',
     'Content-Type': 'application/json',
-  },
+  }),
 };
 
 const requestTimeout = 1000;
@@ -65,10 +65,7 @@ describe('XhrTransport', function () {
             undefined
           );
           assert.strictEqual(request.url, testTransportParameters.url);
-          assert.strictEqual(
-            (request.requestBody as unknown as Blob).type,
-            'application/json'
-          );
+          assert.strictEqual(request.requestBody, testPayload);
           ensureHeadersContain(request.requestHeaders, {
             foo: 'foo-value',
             bar: 'bar-value',