+++ packages/opentelemetry-resources/src/semconv.ts
@@ -1,0 +1,35 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * A local copy of *unstable* semantic conventions.
+ * https://github.com/open-telemetry/opentelemetry-js/blob/main/semantic-conventions/README.md#unstable-semconv
+ */
+export const ATTR_HOST_ARCH = 'host.arch';
+export const ATTR_HOST_ID = 'host.id';
+export const ATTR_HOST_NAME = 'host.name';
+export const ATTR_OS_TYPE = 'os.type';
+export const ATTR_OS_VERSION = 'os.version';
+export const ATTR_PROCESS_COMMAND = 'process.command';
+export const ATTR_PROCESS_COMMAND_ARGS = 'process.command_args';
+export const ATTR_PROCESS_EXECUTABLE_NAME = 'process.executable.name';
+export const ATTR_PROCESS_EXECUTABLE_PATH = 'process.executable.path';
+export const ATTR_PROCESS_OWNER = 'process.owner';
+export const ATTR_PROCESS_PID = 'process.pid';
+export const ATTR_PROCESS_RUNTIME_DESCRIPTION = 'process.runtime.description';
+export const ATTR_PROCESS_RUNTIME_NAME = 'process.runtime.name';
+export const ATTR_PROCESS_RUNTIME_VERSION = 'process.runtime.version';
+export const ATTR_SERVICE_INSTANCE_ID = 'service.instance.id';