+++ packages/opentelemetry-resources/test/Resource.test.ts
@@ -19,10 +19,10 @@ import * as assert from 'assert';
 import { SDK_INFO } from '@opentelemetry/core';
 import { Resource, ResourceAttributes } from '../src';
 import {
-  SEMRESATTRS_SERVICE_NAME,
-  SEMRESATTRS_TELEMETRY_SDK_LANGUAGE,
-  SEMRESATTRS_TELEMETRY_SDK_NAME,
-  SEMRESATTRS_TELEMETRY_SDK_VERSION,
+  ATTR_SERVICE_NAME,
+  ATTR_TELEMETRY_SDK_LANGUAGE,
+  ATTR_TELEMETRY_SDK_NAME,
+  ATTR_TELEMETRY_SDK_VERSION,
 } from '@opentelemetry/semantic-conventions';
 import { describeBrowser, describeNode } from './util';
 import { diag } from '@opentelemetry/api';
@@ -285,19 +285,19 @@ describe('Resource', () => {
     it('should return a default resource', () => {
       const resource = Resource.default();
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_TELEMETRY_SDK_NAME],
-        SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_NAME]
+        resource.attributes[ATTR_TELEMETRY_SDK_NAME],
+        SDK_INFO[ATTR_TELEMETRY_SDK_NAME]
       );
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_TELEMETRY_SDK_LANGUAGE],
-        SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_LANGUAGE]
+        resource.attributes[ATTR_TELEMETRY_SDK_LANGUAGE],
+        SDK_INFO[ATTR_TELEMETRY_SDK_LANGUAGE]
       );
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_TELEMETRY_SDK_VERSION],
-        SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_VERSION]
+        resource.attributes[ATTR_TELEMETRY_SDK_VERSION],
+        SDK_INFO[ATTR_TELEMETRY_SDK_VERSION]
       );
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_SERVICE_NAME],
+        resource.attributes[ATTR_SERVICE_NAME],
         `unknown_service:${process.argv0}`
       );
     });
@@ -307,19 +307,19 @@ describe('Resource', () => {
     it('should return a default resource', () => {
       const resource = Resource.default();
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_TELEMETRY_SDK_NAME],
-        SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_NAME]
+        resource.attributes[ATTR_TELEMETRY_SDK_NAME],
+        SDK_INFO[ATTR_TELEMETRY_SDK_NAME]
       );
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_TELEMETRY_SDK_LANGUAGE],
-        SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_LANGUAGE]
+        resource.attributes[ATTR_TELEMETRY_SDK_LANGUAGE],
+        SDK_INFO[ATTR_TELEMETRY_SDK_LANGUAGE]
       );
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_TELEMETRY_SDK_VERSION],
-        SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_VERSION]
+        resource.attributes[ATTR_TELEMETRY_SDK_VERSION],
+        SDK_INFO[ATTR_TELEMETRY_SDK_VERSION]
       );
       assert.strictEqual(
-        resource.attributes[SEMRESATTRS_SERVICE_NAME],
+        resource.attributes[ATTR_SERVICE_NAME],
         'unknown_service'
       );
     });